requirements.txt
blinker==1.6.2
click==8.1.6
coverage==7.3.0
Flask==2.3.2
Flask-Cors==4.0.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0